apps/ai/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
google-genai>=1.0.0
bullmq-python==0.5.0
pydantic==2.6.1
pydantic-settings==2.1.0
structlog==24.1.0
httpx>=0.27.0
asyncpg>=0.29.0
aiohttp==3.9.3
aioboto3==12.3.0
pymupdf>=1.24.0
pytest==8.0.1
pytest-asyncio==0.23.5
